myweibo_analysis_V2/app/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
PyMySQL==1.0.2
cryptography==3.4.7 